refactor: 다익스트라 라이브러리를 도메인에서 분리 - 전략패턴 사용
